Move Application and services to App namespace
